add hash size argument to engine options
This commit adds a new --hash_size argument to the engine options, which can be used to set a hash for the engine. The argument is an integer and its default value is 16. The new option is used in the function analyze_game along with other options to set the engine options.